requirements.txt
Pillow==9.3.0
tk==0.1.0
tksheet==5.4.1
